Added new test files to test different components that will be integrated into the final code.
testNewWeight.ino is the new method of getting the weight where when someone steps on the scale we take several samples and compare them, making sure that the values are constant (within some error range), and once we get a consistent value, we save that, and wait for the user to step off the scale before exiting the loop.

The second script testStringSearch.ino is just some code to be able to check a character array string and search for a character within it, and return the index.